DCA11Y-1274: Move IncrementalMojoHelper to core
